refactor : XCoordinate 클래스, YCoordinate 클래스 삭제
