--- src/main/java/coordinate/domain/coordinate/XCoordinate.java
@@ -1,7 +1,0 @@
-package coordinate.domain.coordinate;
-
-public class XCoordinate extends Coordinate {
-    public XCoordinate(int value) {
-        super(value);
-    }
-}
