--- src/main/java/coordinate/domain/coordinate/YCoordinate.java
@@ -1,7 +1,0 @@
-package coordinate.domain.coordinate;
-
-public class YCoordinate extends Coordinate {
-    public YCoordinate(int value) {
-        super(value);
-    }
-}
